Specify precision
The f2py package silently assumes a single precision floating point
number as a default. There are numerous references explaining the issue
and a short list is included below.

The solution was to add a module `precis` in the `varPrecision.f90` file
which specifies a parameter `prec` which all values (including so called
Magic Numbers) in the file now respect to force precision.

When the precision is forced to single, the NOLOOP and 2clean versions
of the computation produce exactly the same results. However, when double
precision is forced, there is still a difference between the two versions.


Former-commit-id: a5a509a43ac83c2b1c6384334df27f1963c7fcbf