Remove unnecessary copies.
Significant performance increase. all_hws never changes so there is no
need to create a copy of it every iteration of the loop.

set(all_hws) results in a copy of all_hws (never changes, already a set)
set(sub_hws) results in a copy of sub_hws (not needed because sub_hws is
already a set).
The result of the difference is a set which we copy into a list.

We only need to perform the set difference.