Split LastObs and TimeSlice Reading for looping
LastObs and TimeSlice reading have been combined as part of a
DA read process for non-looping execution. Now, with the
need to separate warm-state reads, which only happen at the
beginning of simulation from forcing reads, which happen throughout
each loop, it was necessary to untangle these two input processes.

As part of that fix, the timeslice read in nhd_io was
modified to assume all globbing has already ocurred in
prior steps and now receives only a list of files to read
and combine. Also, initiated with this commit, the inconsistent use
of "last_obs" and "lastobs" was harmonized to be lastobs in
every case. (This commit should be atomic with regard to the
name update within the affected functions, but a second commit
will see it through to the remainder of the code.)