substitute read_netcdfs for open_mfdataset
There was initially some concern that the asyncio
calls would conflict with sub-thread calls with the
data read and write functions. This commit showed
that a substitution could be made directly, at least
for the read command, and it did change the problem
at the time with the asyncio call -- but ultimately,
once the asynchronous thread pools were being shutdown
properly, there no longer appeared to be a conflict,
and this commit because unnecessary for that purpose.

There is, however, the possibility that the dask
dependencies of the open_mfdataset and write_mfdataset
calls are not allowed on WCOSS -- if that is the case
then something like what is represented here might be
needed, at least as an option, depending on the comparative
performance of the methods.