initialize run_set_iterator
If there is only one run_set, only the initial
forcing call and final model and output calls are
executed with none of the for loop running -- so
we need to initialize the run_set_iterator for that
case.

The regular looping call uses the loop even if there is only one
run_set so this is unique to the slightly different
configuration of the async execution.