test: FilePath and DirectoryPath
Note: FilePath and DirectoryPath build on pydantic's FilePath and
DirectoryPath classes. Specifically, troute.config's version of these
classes unable using a "strict" mode (see commited tests for usage).
The idea of strict mode is that in some contexts it is desirable to
verify if a file or directory path exists. Likewise, in other contexts
that is not desirable. "strict" mode enables the best of both worlds by
effectively toggling if existence is verified.